requirements.txt
chainlit
python-dotenv
httpx
openai 